Added timeout option to fusion_sleep_on() that is used by properties. Leasing/purchasing doesn't fail immediately now if a property is purchased. Instead it fails after a timeout of 20 ms.
This improves DirectFB DRI interoperation significantly as DirectFB doesn't
give up early to access the hardware if an OpenGL application does the rendering
which is most of the time, but depending on the framerate very short periods.
If this period is short enough (say having 40 fps) the timeout will never
expire and no software fallback is used.

Running OpenGL with ~150 fps while scrolling in an IDE which is half transp.
above runs quite smooth.