Implemented lock transfer (halfway) during execution of synchronous calls.
When process A calls process B, the locks held by A are unlocked
and reserved for B's dispatcher thread to lock it. Other threads
would block as if it's still locked.

When B returns the call, the locks are reclaimed. A FUSION_ASSUME will
fail if B did not unlock properly and the lock is hauled to A again.