Fix page faults, incorrect syscall traces in oprofile
The patch replaces the system call frame detection logic to correct the
discovery of system call frames when backtracing in the kernel.  The original
algorithm assumed that if the sp was not a valid kernel stack pointer and
the instruction pointer was within bounds of certain system call handlers,
then a valid system call was found.  This did not take account for other
entry methods into the entry_xx.S common syscall code.

The new algorithm that detects the stack frame in kernel_getsp()
replaces the pc test with:
1) a test for a frame marker in the stack (if supported by variant),
2) a test for an expected value of the stack pointer when in the
   system call frame, and
3) a test to make sure that the code is entered via system call instructions
   rather than other exceptions, by testing the saved pt_regs->trap value.

Also included is a fix to capture one additional user space trace previously
lost in the earlier algorithm.

Signed-off-by: Zhenghui(Sam) Liao <zhenghui.liao@windriver.com>
Signed-off-by: Dave.Lerner<Dave.Lerner@windriver.com>