oprofile: limit sampled cpus
On systems with lots of processors, oprofile may not able to sample all
processors due to bandwidth limitations.  This patch provides a new
oprofilefs control file, sampled_cpus, to select which CPUs get sampled as
a work around.

Reading from the sampled_cpus control file returns a list of CPUs selected
for sampling.  By default, all online CPUs are sampled (example from
Sun Netra T-5440):

  # mount none /mnt/oprofile -t oprofilefs
  # cd /mnt/oprofile
  # cat sampled_cpus
  0-47,72-119

Writing to the sampled_cpus control file selects the CPUs for oprofile to
sample.  The sampled_cpus control file accepts a comma delimited list of
CPUs or CPU range (which must be specified in ascending order).  For example,
the following are valid values for sampled_cpus:

  # echo 0 > sampled_cpus
  # echo 0,1,2,3,4,5,6,7 > sampled_cpus
  # echo 0-7 > sampled_cpus
  # echo 1,4-7 > sampled_cpus
  # echo 0-2,4-7 > sampled_cpus

Processes of interest for profiling should have their CPU affinity set to
include only CPUs that are in sampled_cpus.  The process's CPU affinity could
be set on the command line with the taskset command (part of util-linux)
or with the kernel's cpuset feature.  A detailed description of cpuset
could be found in linux/Documentation/cpusets.txt.

Signed-off-by: Vu Tran <vu.tran@windriver.com>
Signed-off-by: Hong H. Pham <hong.pham@windriver.com>