oprofile: fix unresolved symbols when loading as a module
Use a structure to hold the system call return addresses
which are used by oprofile backtrace during syscall unwinding.
The structure address values are set by code in the x86/kernel
directory, and this will always be built-in when oprofile is enabled.
When oprofile is built as a module, the syscall return addresses
are available through this structure.  The syscall return
addresses are not simply EXPORT'd for security reasons.

Signed-off-by: Dave Lerner <dave.lerner@windriver.com>