MIPS: CGL: Enable kernel snapshots.
This allows taking a snapshot of the state of a running system.  The
snapshot file will be a standard core file so all common tools can be
used to analyze it.  To take a snapshot just echo the filename of the
snapshot to be created into the control file like
"echo /root/snapshot > /sys/ksnap/ksnap".

Depending on the amount of memory in the system and filesystem performance
taking a snapshot may be a slow process; to allow a consistent snapshot
being taken the system will be mostly frozen during the time of the
snapshot.

The implementation is restricted to non-NUMA, non-HIGHMEM systems.

Signed-off-by: Ralf Baechle <ralf.baechle@windriver.com>