p4080ds: implement work-around for erratum CPU6
Do not enable interrupts right before the fourth (and final) call to tlbsync.

Erratum CPU6 says that ME, CE, DE, and EE cannot be enabled when executing a
tlbsync.  The kernel uses tlbsync four times during early boot-up.

A recent change in U-Boot ("ppc/85xx: Disable all async interrupt
sources when we boot") disabled all of these when the kernel boots, but
the kernel turns them back on right before the final call to tlbsync.
However, interrupts dont' need to be enabled at this point; and indeed,
the kernel redundantly enables them again a short time later.  So we
change the code not to enable interrupts right before that last tlbsync
instruction.

Signed-off-by: Timur Tabi <timur@freescale.com>
[Cleanly applied the FSL SDK 2.0.3 patch:
"kernel-2.6.30-p4080ds-implement-work-around-for-erratum-CP.patch"]
Integrated-by: Yuri Nedel <Yuri.Nedel@windriver.com>