sctp: add parameter to enable/disable ECN
SCTP always sends the TLV for ECN in all init chunks, so if
both sides support ECN, SCTP will then set the ECN bits in
the IP header.  If the packets are being routed through a
middlebox that drops packets that have the ECN bits set in
the IP header, SCTP packets with the ECN bits set will be
dropped, and the connection will fail.

To work around situations like this, a new sysctl option is
added: /proc/sys/net/sctp/ecn_enable.  By default it is '1'
meaning all new associations will attempt to negotiate the use
of ECN.  If it is set to '0', new associations will not include
the ECN TLV in init chunks, and it will ignore the ECN TLV if received
in an init chunk.  Changing the state of this option will only
effect new associations, existing associations are unaffected.

Signed-off-by: Greg Moffatt <greg.moffatt@windriver.com>