arm,kgdb: Fix software single step support with a compiled breakpoint
When a breakpoint has been hard compiled into the kernel,
both soft single step, and the continue operation need to
increment the pc by 4, or the cpu will continue to execute
the compiled trap instruction over and over.

The software single step now mimics the behavior of the
continue operation.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>