OMAP3630:DSS2: Updating MAX divider value
commit 27ba6d3282f8053351a94e0ae3518e899c4a1d1c from
the branch origin/OMAPPSP_03.00.01.06 of
git://arago-project.org/git/projects/linux-omap3.git

In DPLL4 M3, M4, M5 and M6 field width has been increased by 1 bit in 3630.
So the max divider value that can be achived will be 32 and not 16.
In 3630 the functional clock is x1 of DPLL4 and not x2 .Hence multiplier 2
is removed.

Signed-off-by: Kishore Y <kishore.y@ti.com>
Integrated-by: Stanley.Miao <stanley.miao@windriver.com>