musb/core: Fix hot plug voltage drop issue
During connection as an A-Device, we may see a short current spikes
causing voltage drop. The software has already got a workaround for
this issue by skipping non-repeat vbuserror, but there are 2 potential
problems, one is that musb may wait for hot plug in A_IDLE or B_IDLE
state, while another is that current solution doesn't support repeatly
hot plug and unplug.

Signed-off-by: Wang Hui <Hui.Wang@windriver.com>