yaffs2: import yaffs2 core support
YAFFS (Yet Another Flash File System) is now in its second generation
and provides a fast robust file system for NAND and NOR Flash. It is
widely used with Linux and RTOSs, in consumer devices. It is dual
licenced under commercial/GPL terms.

This is an import of the latest head of tree from:

  http://www.yaffs.net/:yaffs2-99c62e39578113a798483b5dc6c13a80737163f1

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>