PPC32/kexec: add extra way of waking up secondary CPUs
Some platforms secondary CPUs wakeup algorithms do not fit with the
way kexec holds these CPU while rebooting. Add a configuration flag
that takes care of these cases.

A default shutdown sequence and an very early boot sequence to bring
the secondary CPUs in the new kernel are provided. However, most
probably, a per-platform implementation of KEXEC_POWERPC_SMP_BOOTABLE
is required to wake up the secondary CPUs.

When the secondary CPU goes down in the original kernel, it first
spins in a tight loop in the control code page, which is randomly
allocated, until the copy of the new kernel is done. At that point,
the primary CPU jumps to the new kernel. It then has to bring the
secondary CPU into the second kernel as well since the control code
page has a high chance of being reused by the new kernel before the
primary CPU hits the secondary wakeup code. To achieve this, the
primary CPU writes the address of a new loop in the second kernel
at which the secondary jumps to.

To pass the address of the spin loop variable in the control code
page to the second kernel, register r25 is used: it is callee-saved
by C code that can be called by purgatory and not used by the
purgatory assembly code.

Signed-off-by: Benjamin Walsh <benjamin.walsh@windriver.com>