PPC32/kexec: fix un-reliable handling of SMP crash
Due to the fact that handling a crash takes a different code path
than fast-reboot, and that the previous SMP handling patches did
not take it into account, crash handling was up until now unreliable
on SMP machines.

The generic PPC crash handling uses its own IPI handler to put
secondary CPUs in a tight spin and locks their local interrupts,
before calling the generic kexec code that brings the machine
down. The generic path thus cannot use IPIs to force CPUs onto
the intermediary control code page.

To solve this problem, it figures out if the code path it is coming
from is the crash case or the normal case. If it is the crash case,
it forces CPUs off their first spin loop by writing to the variable
they are spinning on. This variable doubles as a pointer to the kexec
image control structure needed to bring the CPUs into the control page,
instead of introducing another variable. When the CPUs are taken out of
their spin loop, they continue down the generic path with different
reponsibilities depending if they are hardware CPU0 or not (CPU0 has
to be the boot CPU in the new kernel).

Signed-off-by: Benjamin Walsh <benjamin.walsh@windriver.com>