lttng-vmcore-fix
LTTng vmcore fix

Crash dump "LTT_VMCORE" support was broken in many ways. This patch
fixes it. It will be integrated in LTTng 0.114.

Basically, instead of modifying the lost_size in the header, we modify a
new ltt_buf field "commit_seq", for "sequential commits", which counts
the number of consecutive readable bytes in the buffer.

Note that commit_seq should be accessed with a modulo subbuf_size, like
commit_count.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@polymtl.ca>