lttng-relay-lockless-fix-push-reader
LTTng relay lockless fix push reader

lockless relay code needs to push the reader even in the fast path, given this
part is executed our of order.

It is not needed in the locked and irqoff relay because sub-buffer switch
(slow path) is done in order (protected by disabling interrupts).

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@polymtl.ca>