lttng-relay-per-subbuffer-index
lttng relay per subbuffer index

Create per-subbuffer page index in preparation to let the reader exchange the
subbuffer with its own pages to eliminate writer-reader possible data corruption
in flight recorder mode.

Steven Rostedt, with great patience and lenghy explanations, explained to me the
benefit of having an extra "page" (or "subbuffer") owned by the reader for
flight-recorder mode tracing. Credits go to him for getting the idea into my
mind.

Currently, this data corruption is dealt with by checking the return value of
the put_subbuf ioctl. If the reader position has been pushed by the writer, the
reader knows the data it just read was corrupted.

For flight recorder tracing, we always add a supplementary subbuffer to permit
this atomic exchange.

For normal tracing, given the buffer semantic does not allow the writer to
overwrite the reader subbuffer, we never allocate this supplementary subbuffer.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@polymtl.ca>