lttng-relay-lockless-writer-use-noref-flag
LTTng relay lockless writer use noref flag

Update:

Now add cc and cc_sb per-subbuffer commit counters. cc is incremented at each
event with a local_add, then its value is read to check if the subbuffer is
full. We eliminate the local_add_return from the fast path this way, but then
two commits could see the subbuffer being filled.

It's not a problem for delivery detection, but it becomes a problem for
reference drop, because the second writer thread which thinks it must perform
the reference drop would be racing with the reader and the following writer
(after the reader has consumed the data), since it's not protected by the
commit_count vs reserve offset difference.

So, so fix that, we keep a cc_sb counter only incremented with a cmpxchg at each
subbuffer delivery. Only the first thread which detects that cc is full does
increment cc_sb with a cmpxchg, and comparing the previous value of cc_sb makes
sure only one delivery happens per written subbuffer.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@polymtl.ca>