lttng-relay-flush-only-if-reader-present
LTTng relay periodic flush only if reader present

Perform periodic flush only if reader if present for a buffer.
Note that we leave the flush for CPUs entering idle even if no reader is present
because the reader currently has no mean to actively ask for a flush of a remote
buffer. Not sure we'd want to neither, because sending an IPI from a reader to
an idle cpu would wake up the idle cpu and disturb the workload.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@polymtl.ca>