tsc-sync-relax-check
tsc sync relax check

At some point, when booting my 8-core Xeon, 1 out of 8 CPUs seems to be 1700
cycles apart from the others. It might be caused by more frequent SMIs, or by
the CPU just being a bit off.

So let's increase the number of test loops from 10 to 200 (to account for more
frequent SMIs) and relax the accepted range to from 1000 to 3000 cycles.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>