ltt-relay-splice-add-llseek-fop
ltt relay splice add llseek fop

splice.c modifications between 2.6.32 and 2.6.33 :

commit cc56f7de7f00d188c7c4da1e9861581853b9e92f
Author: Changli Gao <xiaosuo@gmail.com>
Date:   Wed Nov 4 09:09:52 2009 +0100

    sendfile(): check f_op.splice_write() rather than f_op.sendpage()

    sendfile(2) was reworked with the splice infrastructure, but it still
    checks f_op.sendpage() instead of f_op.splice_write() wrongly.  Although
    if f_op.sendpage() exists, f_op.splice_write() always exists at the same
    time currently, the assumption will be broken in future silently.  This
    patch also brings a side effect: sendfile(2) can work with any output
    file.  Some security checks related to f_op are added too.

Broke LTTng splice() because it now checks for "NULL" llseek fops. We don't
need, nor provide, any: the offset is used in the splice_read fop directly. So
let's copy the no_llseek kernel function into the splice file (because splice
also checks for !no_llseek) to tell the kernel that we provide a llseek fop
(well, an empty one, but we don't tell that to the kernel !).

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>