lttng-kref-fix
lttng kref fix

Fix handling of concurrent trace session teardown with lttd opening a dentry.
We have to wait for all users of the dentry to be gone (after we removed the
dentry) before we can delete the channel structures.

Problem mostly visible when spawning lttd daemons close to trace teardown
(saving flight recorder channels to disk). An easy way to trigger the bug is to
start/stop flight recorder traces in loop while running hackbench in loops on a
SMP system.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>