PPC32/8572ds/kexec: add KEXEC_SMP_BOOTABLE option
The 8572ds code currently supports two ways of waking up a secondary
CPU. The older uboot is not aware of multiple CPUs and thus leaves
the CPU off and then the kernel starts it; the newer uboot knows
about them and wakes them up and spins them on a tight loop, leaving
the kernel to make them enter the kernel by writing a start address
in the spin loop variable. The kexec-bootable kernel takes an
approach that is similar to the latter, with one more step.

When the secondary CPU goes down in the original kernel, it first
spins in a tight loop in the control code page, which is randomly
allocated, until the copy of the new kernel is done. At that point,
the primary CPU jumps to the new kernel. It then has to bring the
secondary CPU into the second kernel as well since the control code
page has a high chance of being reused by the new kernel before the
primary CPU hits the secondary wakeup code. To achieve this, the
primary CPU writes the address of a new loop in the second kernel
at which the secondary jumps to. When the wakeup code is hit by the
primary, it finally instructs the secondary to run __early_start.

Signed-off-by: Benjamin Walsh <benjamin.walsh@windriver.com>