smc91x: Disable the operation of the SMC91X eeprom
When you read/write the SMC91X eeprom,the operation will be
fail.at the same time,the network will down on the board of
arm_versatile_926ejs.

  $ ethtool -e eth0

The SMC91X eeprom functionality is not core to the BSP,so
we will simple disable the operation to prevent that the
network device don't work.

Signed-off-by: Hongbo Zhong <hongbo.zhong@windriver.com>