yaffs2: fix improper use of const char in yaffs_alias
because vfs_follow_link(struct nameidata *nd, const char *link)
and vfs_readlink(struct dentry *dentry, ...,  const char *link)
need a const char * to pass the link. So it should not be kmalloced
and kfree-ed dynamically.

This will incur to the following results:
	yaffs_follow_link()
	vfs_follow_link()
	link_path_walk()
		|
842             `-> this.name = name;
And this assignment and the related d_inode will be not available
after kfree(alias). so this will trigger no link existence at the
do_last:
1649         if (!(open_flag & O_CREAT)) {
1650                 error = do_lookup(nd, &nd->last, path);
1651                 if (error)
1652                         goto exit;
1653                 error = -ENOENT;
1654                 if (!path->dentry->d_inode)
1655                         goto exit_dput;
			^^^
And so the kernel boot will be failure due to null init.

In yaffs, yaffs_CloneString is actually kmallocing into the local variable
'alias'. So to remove yaffs_CloneString in yaffs_GetSymlinkAlias and the
related kfree meets this fix.

Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>