of/flattree: fix of_flat_dt_is_compatible() to match the full compatible string
commit 883c2cfc8bcc0fd00c5d9f596fb8870f481b5bda upstream
correspond to vendor drop QorIQ-DPAA-SDK-V1-20110609-systembuilder.iso
0040-of-flattree-fix-of_flat_dt_is_compatible-to-match-th.patch

With the current string comparison, a device tree compatible of "foo-bar"
would match as compatible with a driver looking for "foo".  This patch
fixes the function to use the of_compat_cmp() macro so that it does the
right thing on all platforms (If sparc ever uses this code, it will still
want the strncasecmp() behaviour).

Signed-off-by: Stuart Yoder <stuart.yoder@freescale.com>
Signed-off-by: Grant Likely <grant.likely@secretlab.ca>